Étape 1 (Labs 0-2)/requirements.txt
# Updated requirements pour labo 2
pytest>=7.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
psycopg2-binary>=2.9.7
sqlalchemy>=2.0.0
pydantic>=2.0.0
requests>=2.31.0
python-dotenv>=1.0.0
